+++ tests/manual/test.py
@@ -8,7 +8,10 @@ pick_round = 2
 if __name__ == '__main__':
 
     espn = PYESPN('nfl')
+    #espn.load_season_teams_results(season=2024)
     #espn.load_standings(season=season)
+    espn.load_season_league_stat_leaders(season=2024)
+    #espn.load_season_team_stats(season=2024)
     #espn.load_regular_season_schedule(season=season)
     espn.load_seasons_betting_records(season=2024)
     #stats = espn.get_players_historical_stats(player_id=player_id)